\name{bingat-package}
\alias{bingat-package}
\alias{bingat}
\docType{package}
\title{Binary Graph Analysis Tools}

\description{
	This package provides tools for: Calculating the MLE of a set of binary graphs,
	Generating random binary graphs sampled from the Gibb's distribution, and
	Doing regression analysis of a set of binary graphs.
}

\details{
	The following are the types of binary graphs that are accepted:
	\enumerate{
		\item adjMatrix: An entire binary adjacency matrix as a single vector
		\item adjMatrixLT: The upper or lower triangle of a binary adjacency matrix as a single vector
		\item diag: The diagonal vector on a binary adjacency matrix
	}	
}

\author{Terrence Brooks, Berkley Shands, Skye Buckner-Petty, Patricio S. La Rosa, Elena Deych, William D. Shannon}

\keyword{ package }
